refactor: rename github_metric_uses
