+++ app/assets/javascripts/github_metric_uses.coffee
