+++ app/helpers/github_metric_uses_helper.rb
@@ -1,0 +1,2 @@
+module GithubMetricUsesHelper
+end
