--- app/helpers/use_github_metrics_helper.rb
@@ -1,2 +1,0 @@
-module UseGithubMetricsHelper
-end
