+++ app/views/assignments/edit/_general.html.erb
@@ -135,7 +135,7 @@ function microtaskChanged() {
   <!--Github-->
   <tr>
     <td style='padding:5px' id='assignment_use_github' colspan="2">
-      <input type="checkbox" name = "use_github" id="use_github" <%= "checked" if UseGithubMetricsController.exist(@assignment_form.assignment.id) %>></input>
+      <input type="checkbox" name = "use_github" id="use_github" <%= "checked" if GithubMetricUsesController.exist(@assignment_form.assignment.id) %>></input>
       <%= label_tag('use_github', 'Use github metrics?') %>
     </td>
   </tr>
@@ -269,7 +269,7 @@ function microtaskChanged() {
         $('#use_github').change(function () {
             if (jQuery('#use_github').is(':checked')) {
                 jQuery.ajax({
-                    url: '/use_github_metrics/' + <%= @assignment.id %>,
+                    url: '/github_metric_uses/' + <%= @assignment.id %>,
                     method: 'POST',
                     xhrFields: {
                         withCredentials: true
@@ -278,7 +278,7 @@ function microtaskChanged() {
                 })
             } else{
                 jQuery.ajax({
-                    url: '/use_github_metrics/' + <%= @assignment.id %>,
+                    url: '/github_metric_uses/' + <%= @assignment.id %>,
                     method: 'DELETE',
                     xhrFields: {
                         withCredentials: true
