+++ app/views/assignments/list_submissions.html.erb
@@ -16,7 +16,7 @@
 </script>
 <h4><%=@assignment.name%> </h4><h6><%="  Created on:  " + @assignment.created_at.to_s[0..10]%></h6>
 <%#check if use github? %>
-<% if UseGithubMetricsController.exist(@assignment.id) %>
+<% if GithubMetricUsesController.exist(@assignment.id) %>
 <% if session["github_access_token"].nil? %>
   <% topic_identifier, topic_name, users_for_curr_team, participants = get_data_for_list_submissions(@teams.first) %>
   <%= link_to 'Login to Query Github data', { controller: 'metrics', action: 'show', id: participants.first.id } %>
@@ -77,7 +77,7 @@
             <% end %>
             <br>
               <%#check if use github? %>
-              <%if UseGithubMetricsController.exist(@assignment.id) %>
+              <%if GithubMetricUsesController.exist(@assignment.id) %>
                 <% unless Metric.find_by_team_id(team.id).nil? %>
                 <%= link_to 'GitHub metrics', { controller: 'metrics', action: 'show', id: participants.first.id }, target: '_blank' %></td>
                   <% end %>
