--- db/migrate/20211125603108_update_use_github_metric.rb
@@ -1,6 +1,0 @@
-class UpdateUseGithubMetric < ActiveRecord::Migration
-
-  def change
-    add_foreign_key :use_github_metrics, :assignments, column: :assignment_id
-  end
-end
