+++ db/migrate/20211125603111_update_github_metric_uses.rb
@@ -1,0 +1,6 @@
+class UpdateGithubMetricUses < ActiveRecord::Migration
+
+  def change
+    add_foreign_key :github_metric_uses, :assignments, column: :assignment_id
+  end
+end
