+++ spec/features/github_metric_uses_spec.rb
