+++ test/models/github_metric_uses_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-class UseGithubMetricTest < ActiveSupport::TestCase
+class GithubMetricUsesTest < ActiveSupport::TestCase
   # test "the truth" do
   #   assert true
   # end
